+++ src/rect.rs
@@ -2,14 +2,14 @@
 #[derive(Copy, Clone)]
 pub struct Rect {
     // TODO: Make these properties private
-    pub x: u32,
-    pub y: u32,
-    pub w: u32,
-    pub h: u32,
+    pub x: u16,
+    pub y: u16,
+    pub w: u16,
+    pub h: u16,
 }
 
 impl Rect {
-    pub fn from(x: u32, y: u32, w: u32, h: u32) -> Rect {
+    pub fn from(x: u16, y: u16, w: u16, h: u16) -> Rect {
         Rect { x, y, w, h }
     }
 }
